% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable.R
\name{handsontableOutput}
\alias{handsontableOutput}
\title{Widget output function for use in Shiny}
\usage{
handsontableOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable name}

\item{width, height}{Widget dimensions}
}
\description{
Widget output function for use in Shiny
}
