% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_cols}
\alias{hot_cols}
\title{Configure Handsontable Columns}
\usage{
hot_cols(
  hot,
  colWidths = NULL,
  fixedColumnsLeft = 0,
  manualColumnResize = TRUE,
  manualColumnMove = FALSE,
  ...
)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{colWidths}{Numeric vector or single value for column widths}

\item{fixedColumnsLeft}{Number of columns to freeze on the left}

\item{manualColumnResize}{Logical, enable manual column resizing}

\item{manualColumnMove}{Logical, enable manual column reordering}

\item{...}{Additional column configuration options}
}
\value{
Modified handsontable widget
}
\description{
Configure Handsontable Columns
}
