% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_row}
\alias{hot_row}
\title{Configure Row Properties}
\usage{
hot_row(hot, row, readOnly = NULL, ...)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{row}{Row index to configure}

\item{readOnly}{Logical, make row read-only}

\item{...}{Additional row configuration options}
}
\value{
Modified handsontable widget
}
\description{
Configure Row Properties
}
