% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_rows}
\alias{hot_rows}
\title{Configure Handsontable Rows}
\usage{
hot_rows(
  hot,
  fixedRowsTop = 0,
  fixedRowsBottom = 0,
  manualRowResize = TRUE,
  manualRowMove = FALSE,
  ...
)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{fixedRowsTop}{Number of rows to freeze at the top}

\item{fixedRowsBottom}{Number of rows to freeze at the bottom}

\item{manualRowResize}{Logical, enable manual row resizing}

\item{manualRowMove}{Logical, enable manual row reordering}

\item{...}{Additional row configuration options}
}
\value{
Modified handsontable widget
}
\description{
Configure Handsontable Rows
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    titlePanel("Row Configuration"),
    handsontableOutput("table")
  )

  server <- function(input, output, session) {
    output$table <- renderHandsontable({
      handsontable(mtcars[1:15, 1:5]) |>
        hot_rows(
          fixedRowsTop = 2, manualRowMove = TRUE,
          manualRowResize = TRUE
        )
    })
  }

  shinyApp(ui, server)
}

}
