% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_table}
\alias{hot_table}
\title{Configure General Table Options}
\usage{
hot_table(
  hot,
  contextMenu = NULL,
  dropdownMenu = NULL,
  filters = NULL,
  manualColumnSorting = NULL,
  multiColumnSorting = NULL,
  search = NULL,
  undo = NULL,
  outsideClickDeselects = NULL,
  fillHandle = NULL,
  themeName = NULL,
  ...
)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{contextMenu}{Logical or list, enable/configure context menu}

\item{dropdownMenu}{Logical or list, enable/configure dropdown menu}

\item{filters}{Logical, enable column filters}

\item{manualColumnSorting}{Logical, enable manual column sorting}

\item{multiColumnSorting}{Logical, enable multi-column sorting}

\item{search}{Logical, enable search functionality}

\item{undo}{Logical, enable undo/redo functionality}

\item{outsideClickDeselects}{Logical, deselect on outside click}

\item{fillHandle}{Logical or object, configure fill handle}

\item{themeName}{Character, theme name (e.g., "ht-theme-main-dark-auto")}

\item{...}{Additional table configuration options}
}
\value{
Modified handsontable widget
}
\description{
Configure General Table Options
}
