% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_validate}
\alias{hot_validate}
\title{Configure Data Validation}
\usage{
hot_validate(
  hot,
  cols,
  type = c("numeric", "list", "regexp"),
  source = NULL,
  pattern = NULL,
  min = NULL,
  max = NULL,
  allowInvalid = TRUE,
  strict = FALSE,
  ...
)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{cols}{Column indices to apply validation (numeric vector)}

\item{type}{Validation type: "numeric", "list", "regexp"}

\item{source}{For "list" type, vector of allowed values}

\item{pattern}{For "regexp" type, regular expression pattern}

\item{min, max}{For "numeric" type, minimum and maximum values}

\item{allowInvalid}{Logical, allow invalid values to be entered}

\item{strict}{Logical, strict validation mode}

\item{...}{Additional validation options}
}
\value{
Modified handsontable widget
}
\description{
Configure Data Validation
}
