% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable.R
\name{renderHandsontable}
\alias{renderHandsontable}
\title{Widget render function for use in Shiny}
\usage{
renderHandsontable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression that generates a handsontable}

\item{env}{Environment in which to evaluate expression}

\item{quoted}{Logical indicating whether expression is quoted}
}
\description{
Widget render function for use in Shiny
}
