% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{set_data}
\alias{set_data}
\title{Update Cell Data Programmatically}
\usage{
set_data(id, row, col, val, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Character. The output ID of the handsontable widget}

\item{row}{Numeric. Row index (1-based)}

\item{col}{Numeric. Column index (1-based)}

\item{val}{Any. Value to set in the cell}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effects)
}
\description{
Update Cell Data Programmatically
}
