% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_context_menu}
\alias{hot_context_menu}
\title{Configure Context Menu}
\usage{
hot_context_menu(
  hot,
  allowRowEdit = TRUE,
  allowColEdit = TRUE,
  customOpts = NULL,
  ...
)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{allowRowEdit}{Logical, allow row operations in context menu}

\item{allowColEdit}{Logical, allow column operations in context menu}

\item{customOpts}{List of custom context menu options}

\item{...}{Additional context menu options}
}
\value{
Modified handsontable widget
}
\description{
Configure Context Menu
}
\examples{
if (interactive()) {
library(shiny)

ui <- fluidPage(
  titlePanel("Context Menu Configuration"),
  handsontableOutput("table")
)

server <- function(input, output, session) {
  output$table <- renderHandsontable({
    handsontable(mtcars[1:8, 1:5]) |>
      hot_context_menu(allowRowEdit = TRUE, allowColEdit = TRUE)
  })
}

shinyApp(ui, server)
}

}
