% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{hot_row}
\alias{hot_row}
\title{Configure Row Properties}
\usage{
hot_row(hot, row, readOnly = NULL, ...)
}
\arguments{
\item{hot}{A handsontable widget object}

\item{row}{Integer vector. Row indices to configure}

\item{readOnly}{Logical, make row read-only}

\item{...}{Additional row configuration options}
}
\value{
Modified handsontable widget
}
\description{
Configure Row Properties
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    titlePanel("Row Configuration"),
    p("Rows 1, 3, and 5 are read-only"),
    handsontableOutput("table")
  )

  server <- function(input, output, session) {
    output$table <- renderHandsontable({
      hot_table <- handsontable(iris[1:8, ])
      for (i in c(1, 3, 5)) {
        hot_table <- hot_table |> hot_row(row = i, readOnly = TRUE)
      }
      hot_table
    })
  }

  shinyApp(ui, server)
}

}
