% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable.R
\name{renderHandsontable}
\alias{renderHandsontable}
\title{Widget render function for use in Shiny}
\usage{
renderHandsontable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression that generates a handsontable}

\item{env}{Environment in which to evaluate expression}

\item{quoted}{Logical indicating whether expression is quoted}
}
\description{
Widget render function for use in Shiny
}
\examples{
if (interactive()) {
library(shiny)

ui <- fluidPage(
  titlePanel("Interactive Handsontable"),
  handsontableOutput("editable_table")
)

server <- function(input, output, session) {
  output$editable_table <- renderHandsontable({
    handsontable(iris[1:10, ], readOnly = FALSE) |>
      hot_cols(colWidths = 120) |>
      hot_table(contextMenu = TRUE)
  })
}

shinyApp(ui, server)
}

}
