% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_functions.R
\name{set_data}
\alias{set_data}
\title{Update Cell Data Programmatically}
\usage{
set_data(id, row, col, val, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Character. The output ID of the handsontable widget}

\item{row}{Numeric. Row index (1-based)}

\item{col}{Numeric. Column index (1-based)}

\item{val}{Any. Value to set in the cell}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effects)
}
\description{
Update Cell Data Programmatically
}
\examples{
if (interactive()) {
library(shiny)

ui <- fluidPage(
  titlePanel("Programmatic Cell Updates"),
  handsontableOutput("my_table"),
  br(),
  actionButton("update_btn", "Update Cell (1,1)")
)

server <- function(input, output, session) {
  output$my_table <- renderHandsontable({
    handsontable(data.frame(
      A = c(1, 2, 3),
      B = c("a", "b", "c")
    ))
  })

  observeEvent(input$update_btn, {
    set_data("my_table", row = 1, col = 1, val = runif(1, 100, 999))
  })
}

shinyApp(ui, server)
}

}
